Tests in user-storage-simple quickstart are failing in main branch closes #567

+++ extension/user-storage-simple/src/test/java/org/keycloak/quickstart/util/StorageManager.java
@@ -40,6 +40,7 @@ public class StorageManager {
     }
 
     public static void deleteStorage() {
+        if (file == null) return;
         try {
             Files.deleteIfExists(file);
         } catch (IOException e) {
